+++ activerecord/CHANGELOG.md
@@ -1,5 +1,18 @@
+*   .unscope and .except now act on the default scope. For example:
+
+      class Project < ActiveRecord::Base
+        default_scope -> { where trashed: false }
+      end
+
+      Project.unscope(where: :trashed).to_sql
+      #=> "SELECT \"projects\".* FROM \"projects\""
+      Project.except(:where).to_sql
+      #=> "SELECT \"projects\".* FROM \"projects\""
+
+    *Dave Jachimiak*
+
 *   Flatten merged join_values before building the joins.
-   
+
     While joining_values special treatment is given to string values.
     By flattening the array it ensures that string values are detected
     as strings and not arrays.